+++ CHANGES.rst
@@ -7,6 +7,7 @@ Change log
 
 - Fix ``SIGCHLD``/``wait`` raise condition associated with the
   ``start-test-program`` option.
+  For details see `#33 <https://github.com/zopefoundation/zdaemon/pull/33>`_.
 
 
 5.0 (2023-05-24)